+++ components/Content/LeftColumn/Categories/index.styl
@@ -1,0 +1,28 @@
+// 1u is a 8px grid unit. 1u === 8px
+$headers = #686868
+$textColor = #979797
+
+.root
+  justify-content flex-start
+
+.header
+  color $headers
+  font-size 2.1u
+  margin-top 5u
+  
+  &.first
+    margin-top 0u
+    
+
+/* CATEGORIES */
+.categoriesBlock
+  margin-top 2u
+
+.cat
+  flex-direction row
+  justify-content space-between
+
+.text
+  color $textColor
+  font-weight 400
+  line-height 4u
