+++ components/Content/LeftColumn/Colors/index.styl
@@ -1,0 +1,31 @@
+// 1u is a 8px grid unit. 1u === 8px
+$headers = #686868
+$textColor = #979797
+
+.root
+  justify-content flex-start
+
+.header
+  color $headers
+  font-size 2.1u
+  margin-top 5u
+  
+  &.first
+    margin-top 0u
+  
+/* COLORS */
+.colorsBlock
+  flex-direction row
+  margin-top 3u
+
+.colors
+  margin-left 2u
+  
+  &.first
+    margin-left 0u
+
+.colorBtn
+  height 20px
+  width 20px
+  border-radius 9999px
+
