+++ components/Content/LeftColumn/Products/index.styl
@@ -1,0 +1,55 @@
+// 1u is a 8px grid unit. 1u === 8px
+$headers = #686868
+$textColor = #979797
+
+.root
+  justify-content flex-start
+
+.header
+  color $headers
+  font-size 2.1u
+  margin-top 5u
+  
+  &.first
+    margin-top 0u
+
+/* RELATED PRODUCT */
+.relatedBlock
+  margin-top 3u 
+  
+.productInfo
+  flex-direction row
+  margin-top 3u
+
+  &.first
+    margin-top 0u
+
+.imageProduct
+  width 10u
+  height 10u
+
+.about
+  margin-left 2u
+
+.name
+  font-size 1.5u
+  color $headers
+
+.price
+  font-size 1.5u
+  color $headers
+  margin-top 2u
+
+.starsBlock
+  margin-top 2u
+  flex-direction row
+
+.stars
+  margin-left 6px
+
+  &.first
+    margin-left 0px
+
+.imageStar
+  width 10px
+  height 10px
