+++ components/Content/LeftColumn/Sizes/index.styl
@@ -1,0 +1,30 @@
+// 1u is a 8px grid unit. 1u === 8px
+$headers = #686868
+$textColor = #979797
+
+.root
+  justify-content flex-start
+
+.header
+  color $headers
+  font-size 2.1u
+  margin-top 5u
+  
+  &.first
+    margin-top 0u
+
+/* SIZE */
+.sizeBlock
+  flex-direction row
+  margin-top 3u
+
+.size
+  margin-left 2u
+  
+  &.first
+    margin-left 0u
+
+.text
+  color $textColor
+  font-weight 400
+  line-height 4u
