+++ components/Footer/index.styl
@@ -123,6 +123,9 @@ $colorWhite = #ffffff
   margin-top 1.5u
   flex-direction row
 
+  &.first
+    margin-top 0u
+
 .value
   color $textColor 
   font-size 1.6u
